fix empty template_uuid string
template_uuid is tested using !=nil while it is set by default to
an empty string. this led to the empty string being used
as value in the json send to the server.

this adds a test (like for sshkey_uuid) if template_uuid is the
empty string and sets it to nil if so, preventing it to be
considered in the json body.